refactor: refactored file names, types and directories
