+++ src/App.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import Header from './components/margins/Header'
-import Footer from './components/margins/Footer'
-import PageHome from './components/pages/PageHome'
-import PageBrowser from './components/pages/PageBrowser'
-import PageMyMovies from './components/pages/PageMyMovies'
-import PageMovie from './components/pages/PageMovie'
+import Header from './components/Margin/Header'
+import Footer from './components/Margin/Footer'
+import PageHome from './components/Pages/PageHome'
+import PageBrowser from './components/Pages/PageBrowser'
+import PageMyMovies from './components/Pages/PageMyMovies'
+import PageMovie from './components/Pages/PageMovie'
 import './resources/fontawesome/css/fontawesome.css'
 import './resources/fontawesome/css/solid.css'
 import ScrollToTop from './components/utils/ScrollToTop'
