+++ src/components/Browse/SearchMovieCard.tsx
