+++ src/components/Card/Thumbnail.tsx
