+++ src/components/Expanded-Movie/ExpandedThumbnail.tsx
