+++ src/components/utils/Burger/hideBurger.ts
